+++ dart/utils/FileInfoWorld.cpp
@@ -112,7 +112,7 @@ bool FileInfoWorld::loadFile(const char* _fName)
     }
 
     state.resize(tempState.size());
-    for (int j = 0; j < tempState.size(); j++)
+    for (size_t j = 0; j < tempState.size(); j++)
       state[j] = tempState[j];
     mRecord->addState(state);
     tempState.clear();
